Trait band::pallet::Config

pub trait Config<I: 'static = ()>: Config {
    type Symbol: Parameter + Ord;
    type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>;
    type WeightInfo: WeightInfo;
    type OnNewSymbolsRelayedHook: OnNewSymbolsRelayed<Self::Symbol>;
    type GetBandRateStalePeriod: Get<<<Self as Config<I>>::Time as Time>::Moment>;
    type GetBandRateStaleBlockPeriod: Get<Self::BlockNumber>;
    type MaxRelaySymbols: Get<u32>;
    type Time: Time;
    type OnSymbolDisabledHook: OnSymbolDisabled<Self::Symbol>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Symbol: Parameter + Ord

Type of the symbol to be relayed.

type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>

Event type of this pallet.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type OnNewSymbolsRelayedHook: OnNewSymbolsRelayed<Self::Symbol>

Hook which is being executed when some new symbols were relayed

type GetBandRateStalePeriod: Get<<<Self as Config<I>>::Time as Time>::Moment>

Rate expiration period in seconds.

type GetBandRateStaleBlockPeriod: Get<Self::BlockNumber>

Rate expiration period in blocks

type MaxRelaySymbols: Get<u32>

Maximum number of symbols that can be relayed within a single call.

type Time: Time

Time used for checking if rate expired

type OnSymbolDisabledHook: OnSymbolDisabled<Self::Symbol>

Hook which is being executed when some symbol must be disabled